# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mercadeiro API
  description: >
    API do Mercadeiro — app de lista de compras domiciliar (Despensa → Carrinho →
    Finalizar Compra → Analítico). Documenta os endpoints da Fase 1 (API core).
    Regras de negócio referenciadas (RN-xx) e casos de uso (UC-xx) estão detalhados
    em `mercadeiro-regras-de-negocio.md`, na raiz do monorepo.


    **Autenticação:** todas as rotas exigem `Authorization: Bearer <token>`,
    exceto `/health`. Token único de aplicação (app single-user, sem contas —
    ver decisão em `mercadeiro-regras-de-negocio.md` §5.8), configurado em
    `API_BEARER_TOKEN` no `.env` do backend. Requisições sem token válido
    recebem `401 Unauthorized`.
  version: 0.1.0
servers:
  - url: http://localhost:3333
    description: Ambiente local de desenvolvimento
security:
  - bearerAuth: []
...
paths:
...
  /categorias:
    get:
      tags: [Categorias]
      summary: Lista todas as categorias, ordenadas por `ordem`
      operationId: listarCategorias
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Categoria"
    post:
      tags: [Categorias]
      summary: Cria uma nova categoria
      operationId: criarCategoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
                  minLength: 1
                ordem:
                  type: integer
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "201":
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"
        "400":
          $ref: "#/components/responses/BadRequest"
  /categorias/{id}:
    parameters:
      - $ref: "#/components/parameters/CategoriaId"
    get:
      tags: [Categorias]
      summary: Busca uma categoria por id
      operationId: buscarCategoria
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Categoria encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Categorias]
      summary: Atualiza nome e/ou ordem de uma categoria
      operationId: atualizarCategoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  minLength: 1
                ordem:
                  type: integer
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Categoria atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Categorias]
      summary: Remove uma categoria sem produtos vinculados
      operationId: removerCategoria
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "204":
          description: Categoria removida
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Categoria possui produtos vinculados e não pode ser removida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    CategoriaId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Token único de aplicação (API_BEARER_TOKEN no .env do backend).
        Exigido em todas as rotas, exceto /health.
  responses:
...
    Unauthorized:
      description: Token de autenticação ausente ou inválido
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          description: >
            Detalhes adicionais, quando aplicável — formato varia por origem:
            array de issues do Zod em erros de validação, objeto de metadados
            do Prisma em conflitos de unicidade, ou ausente.
...
    Categoria:
      type: object
      required: [id, nome, ordem, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        ordem:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...